python/ocr/requirements.txt
alembic==1.13.2
boto3==1.20.9
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7 
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
psycopg2==2.9.10
redis==5.0.1
requests==2.32.3
pytesseract==0.3.10
SQLAlchemy==2.0.31
PyMuPDF==1.25.3
python-dotenv==1.0.1
anthropic==0.49.0